V1.0
This is the completed version of my we app that has been cleaned up and with some comments added. The javascript has been switched from if else chains to a switch and has some minor improvements. Furthmore the table has een remodeled and a donate paypal and bitcoin section has been added as the wesite will be ad free